shared magic energy examine
